feat: Provide pre-commit-which-hook-entry script
This script is useful for cases when text editor plugin is not able to
use run hook entry script, but it allows to provide full path to the
tool, such as eslint or prettier.

Also mark the script as dangerous to use cause of fact that pre-commit
internals does not intend to be used in external scripts.

Fixes: #3
Fixes: #4
Fixes: #8